chore: Add deployment workflow
A new deployment workflow has been added to the repository. This workflow is triggered on push events to the master branch and deploys the application to a server. The workflow includes steps to checkout the code, copy the SSH key, add SSH known hosts, and deploy the application using Docker Compose.

Co-authored-by: luisignaciocc