chore: Add lo-fi chill hip hop instrumental structure generation parameters
This commit adds the necessary code changes to include the parameters for generating a lo-fi chill hip hop instrumental structure. The code includes enums for different process types, as well as a type for the process parameters. It also includes a specific set of parameters for generating lo-fi instrumental structures. These parameters define the messages to be used in the OpenAI chat completions for generating the structure, as well as the video description and tags. Additionally, a function is provided to generate a title based on the input title. This commit is part of the ongoing development of the application.